Add helm Chart for adding PostgreSQL for Dawson app
This is the beginning of implementing a work app called "Dawson"
into a Kubernetes cluster.

Dawson has a few services that it needs to run, such as PostgreSQL.
I will be using helm to accomplish this, since I do not want to write
a bunch of extra .yml files if not needed, I trust bitnami with that.